mav_interface_msgs ported to ROS2
ROS2 messages are broken. First of all, there are no enums anymore, so now we will have to set integers
in the code and manually do the enums ourselves (fun stuff). Also, the members of a ROS2 message can not be
written in camelCase, since it will throw some error (nonesense). This is the reason why I have to switch to snake case
naming, so that these can be compiled.

As a ToDo, I still have to figure out a way of compiling the functions that we had before.